Update of the Project_Information Page and rename Information.tsx to Project_Information.tsx
